+++ examples/data-sources/meraki_appliance_network_security_intrusion/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_network_security_intrusion" "example" {
+  network_id = "L_123456"
+}