+++ examples/resources/meraki_appliance_network_security_intrusion/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_network_security_intrusion.example "<network_id>"