+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add `meraki_appliance_rf_profiles` data source
 - BREAKING CHANGE: Rename `per_ssid_settingsXX_*` attributes of `meraki_wireless_rf_profile` resource and data sources to `per_ssid_settings_XX_*`
 - Add `meraki_appliance_sdwan_internet_policies` resource
+- Add `meraki_appliance_network_security_intrusion` resource and data source
 
 ## 0.1.1